deps: Use PR branch (#328) for @react-native-community/cameraroll.
Just while we wait for the code in this package to be error-free
when exact_by_default is enabled [1]. We plan to enable it for the
RN v0.64 upgrade (#4426). The PR is
  https://github.com/react-native-cameraroll/react-native-cameraroll/pull/328.

It's very possible that some of those object types would be better
as exact instead of inexact. But:

1. Just getting the library to pass the implicit-inexact-object lint
   is enough to unblock switching on exact_by_default (for us and
   for other projects that want to do that too).

2. I guess we could go through and decide what should stay inexact
   and what should be made exact. But then each of those choices
   would have to be approved, and some of them would probably be
   breaking changes that we'd need to write changelog entries for.
   (Not ones that would affect an app at runtime -- but still, apps
   that use Flow would predictably have to change their code to
   adapt.)

So instead, the PR just makes all the implicit inexact object types
explicitly inexact.  It's based directly on v4.0.4, the version we've
been using, so it doesn't pull in any other changes.

[1] https://medium.com/flow-type/how-to-upgrade-to-exact-by-default-object-type-syntax-7aa44b4d08ab