deps: Add fbjs-scripts, for the benefit of jest-expo when we upgrade RN.
`jest-expo` has a bug where it imports `fbjs-scripts` library but
doesn't declare it as a dependency. This hasn't caused any issues so
far because React Native has been pulling in `fbjs-scripts`, and
`jest-expo` uses it from there.

But RN v0.64 stops pulling in `fbjs-scripts`, so things fail [1].

So, until we get a version of `jest-expo` that declares
`fbjs-scripts` as a dependency, or (preferably) consumes the code
from its new home at @jest/create-cache-key-function [2] (and
hopefully declares *that* as a dependency), we should keep
`fbjs-scripts` as our own direct dependency. Since this is meant to
be a temporary state, and to keep down complexity, just pin to the
exact version that RN pulls in in RN v0.63, which is 1.2.0.

[1] https://github.com/expo/expo/issues/13886#issuecomment-891905293
[2] See facebook/react-native@fd9787ecc.