android build: Upgrade Android Gradle Plugin to v4.1.0.
We decided to postpone taking 4.0.1 or later until after the RN
v0.64 upgrade. That's because someone reported an issue with
4.0.1 [1], and a supposed fix landed in
facebook/react-native@53f55001a, which was a change to RN's internal
code released in v0.64.

Release notes here:
  https://developer.android.com/studio/releases/gradle-plugin#4-1-0

Quite a few major changes announced, hopefully they're all good
ones. Our tests pass, and a debug build on Android built and ran
fine for me.

Part of the RN v0.63 -> v0.64 changes to the template app,
corresponding to:

  facebook/react-native@cf8368f20
  facebook/react-native@553fb8b28
  facebook/react-native@dfa9db49e

See discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Pin.20to.20explicit.20NDK.20version.3F/near/1210089.

[1] https://github.com/facebook/react-native/pull/29013#issuecomment-708079213