changelog: Tweak changelog for v27.170.
Issue #4909 is one that we're not aware of anyone seeing in practice;
it was spotted while reading the source code.  So we can leave it out
of the user highlights.

For issue #3472, the conditions for when it applies are potentially
confusing, but flip them around in a way that's hopefully clearer.
Also go back and add a reference to where we previously fixed its
most visible symptom (that even the @-mentions view didn't have
these messages.)

Also make a couple of small stylistic tweaks:
 * simplify for users by not mentioning "PR ";
 * issue/PR number follows sentence-ending punctuation -- just like
   a footnote marker -- so that the reader who's not interested in
   the extra information can tune it out and see the sentence close
   normally.

In the developer highlights, where we get wordier, blank lines
between items can be helpful for the reader to scan and visually
parse.  Also make some small wording and formatting tweaks there.