api errors: Make the invalid-JSON case a bit more explicit
As it stands, if we didn't get a response from the server or its
body was invalid JSON (probably by being empty), then the `response`
property just disappears in the Sentry breadcrumb.  That can be a
bit confusing when matching up the Sentry data with the source code.

Instead, say explicitly what we got.