ZulipAsyncStorage: Document and assert JSON requirement
We've always passed JSON-encoded data in the one place we actually
use this storage, inside the redux-persist code; and our `getItem`
has relied on that assumption for a long time.

But our tests for this module have been passing random strings
instead.  When those happen to start with a `z`, that collides
with our compression header, and the test flakes.

Document that requirement on this module's actual interface, and
add some debug asserts to enforce it.  That then makes the flakes
much easier to reproduce; fix those tests to conform to the
interface's requirements.

Fixes: #4994