webview html: Skip checking isSameRecipient when we know the answer
This change is NFC except for performance: all we're doing is
inlining diffRecipient into the one place it's used.

Because `&&` short-circuits, this means we'll skip calling
isSameRecipient when we're in a conversation narrow, where we're
not going to show recipient headers anyway.  That's good because
that function does rather more computation than one would like,
particularly for PMs.