event [nfc]: Unpack and TODO-comment account-switch logic in poll loop
The small refactor here unpacks the two steps that the now-deprecated
tryGetAuth was combining: getting the state for the active account
(and rejecting if there's none), and then getting the auth (and
rejecting if we're logged out.)

While staring at this, also expand the comments on what this logic is
doing.  That includes adding some TODO comments, because it doesn't
seem like the code quite accomplishes what it's trying to; the
comments refer to one existing issue and one I just filed for this.