account: Mark getAuth as per-account
I've just read through all the callers, and confirmed that they're
all per-account.  Which is what one would hope, given that this
selector's interface was already that it'll throw an exception if
there isn't an active account.

As a bonus, an untruthful error message gets fixed.  We've been
reporting "Active account not logged in" not only in the case
where indeed the active account is not logged in, but also in the
case where there is no active account at all.

This is NFC other than that change in error message.