notif [nfc]: Mark tryStopNotifications as operating per-account
This is invoked from the main app UI, so its caller knows which
account it's meant to operate on.

Marking as per-account is expressed only through jsdoc, as we don't
yet have types to distinguish thunk actions as per-account vs. global.
(That's coming later.)  While we're documenting this function, write
down some caveats too.

Having marked it as per-account, we can move it off the deprecated
getActiveAccount to the explicitly per-account version getAccount.