msglist [nfc]: Give just one name to "message peer" and "HTML piece"
Namely, "message-list element".

When we implement a message-list diffing algorithm, we won't be able
to diff list A with list B without a clear idea of what an item in
each list actually is. (Is it a recipient header plus all the
following messages it describes? Is it a recipient header plus one
message? And so on.)

Those items (currently) are:

- Recipient header
- Date separator
- Message

In the React Native side, we've been calling these "HTML pieces", as
in `htmlPieceDescriptors`.

In the WebView, we've been calling them "message peers".

Give them a single name, so it's clearer that they represent a
single concept in our future message-list diffing algorithm.