messageListElementHtml [nfc]: Rename some private helpers
In particular:

- messageAsHtml -> message
- messageHeaderAsHtml -> header
- timeRowAsHtml -> time

The new names follow the `type` property on `MessageListElement`
objects, and they're intended to make it clearer that they are, in
fact, private helpers of `messageListElementHtml`.

For our future message-list diffing algorithm, there will need to be
a 1:1 correspondence between a MessageListElement object on the RN
side and a specific DOM node in the WebView. Emphasizing that these
are private helpers of `messageListElementHtml` should help a bit
with that.

Also give the helpers a jsdoc to reinforce the point.