redux types: Stop letting GlobalState pose as PerAccountState
Well, without fixmes, anyway.  Most of which (just a handful)
we've added in the preceding few commits.

In this commit we add two fixmes as plain fixme comments, rather
than calls to nice explicit functions.  One is adjacent to an
existing dubJointState call and will naturally go away when that
one does; and another fixme in reducers.js, which will be easier to
resolve when we do some re-organizing of that code to distinguish
per-account vs. global reducers, coming soon.