redux types: Add GlobalDispatch and GlobalThunkAction
For now the plain Dispatch still provides a GlobalState (which in turn
can still for now be used as a PerAccountState).  With these separate
names we'll be able to start distinguishing thunk actions that really
are meant to be global from those that are per-account.  Then we'll
be able to tighten the plain ThunkAction type (and Dispatch, which is
tied to it) to provide only a PerAccountState.