persist: Retry failed writes on next update
As long as the multiSet always succeeds (and the admonition on
when not to call _resetLastWrittenState is followed), this should
be NFC: on each call to writeOnce, outstandingKeys will be empty
and lastWrittenState will have the same value it would have had
without this change.

The difference comes when a multiSet fails.  With the old code,
we'd have lastWrittenState still reflect the previous state, so
we'd *usually* retry writing all the keys we wrote last time...
but if the new-new value at a key happens to equal the old value,
we wouldn't.  For example if for some key the value was
  Immutable.List()
and then became
  Immutable.List([1])
and then back to
  Immutable.List() ,
and if the write for the first update didn't succeed, then on the
second update we'd see that the new value was Immutable.List() --
which is ===-equal to any other (empty) Immutable.List() -- and
wouldn't try to write it out.

Because a write can always fail (at getting success back to us)
*after* the actual storage has been updated with the new value,
this means the stored value in that example could stay at the
now-stale Immutable.List([1]) indefinitely.  Previous discussion:
  https://github.com/zulip/zulip-mobile/pull/4694#discussion_r691727392

Instead, we now track keys we've attempted a write for that we
don't know succeeded, and always retry those on the next update.

This also lets us redefine lastWrittenState to refer to the last
*attempted* write, rather than the last successful write.  That
means it gets updated atomically with where we compare against it,
which simplifies reasoning about its behavior and lets us remove a
caveat from this module's interface.