app bar [nfc]: Split out TopTabButton from NavButton
Logically these are really two different contexts, where we
potentially want different layout behavior: an app bar on one
hand, and a sort of inverted bottom-navigation (like we have on
`HomeScreen`) on the other.

At present they have the same implementation.  We want to adjust
the layout of the buttons in the app bar, which will involve
changes here, but would rather keep the complexity down by not
worrying simultaneously about how it affects `HomeScreen`.
So, this makes a good time to split these up.

The new TopTabButton.js is a copy of NavButton.js, just with the
components renamed.  For example, try:
  $ git diff --no-index src/nav/{Nav,TopTab}Button.js