app bar [nfc]: Fold away NavButtonGeneral
With TopTabButton, we use the flexibility that TopTabButtonGeneral
gives us to provide an icon that's more than a plain `Icon`.

But with its twin NavButton -- which we're about to adjust the
layout of to better conform to the desired spec, making it not
so much of a twin -- we don't.

So, inline NavButtonGeneral away, to simplify this module's interface
and make it easier to reason about changes to this code.