app bar: Fix height, especially when no back button
These app bars -- all our app bars outside the message list, which
uses ChatNavBar -- didn't have anything to actually specify how tall
they should be.

When the back button is present, it serves as a strut to keep the app
bar at least as tall as the button is, which gets us pretty close:
just a fraction of a pixel off in my testing on an Android 11
emulator.  (In fact I don't understand why it isn't exactly right:
`NavButtonGeneral` sets the height to NAVBAR_SIZE.)

But when the back button is not present, the app bar would collapse to
the height of the text, which is typically much shorter.

Fix both cases by simply specifying the intended height, as a minimum.

We actually used to have explicit heights like these, as simply
`height: NAVBAR_SIZE`, and removed them a little while ago in #4443
in commits 635b13a39 and a9ab3fc21.  The stated reason was to better
handle large font sizes; it also simplified rearranging things for
getting these `SafeAreaView`s in place.  But those have settled down,
and by using `minHeight` there's no trouble if the user has a very
large font-size setting.  (Also I don't think we noticed at the time
that this had a large visible effect in the no-back-button case; if
we had, we'd have done something about it then.)

In the case of `ChatNavBar`, we do specify the height, and ideally
we'd say `minHeight` instead there too.  But that causes hilariously
wrong results, presumably due to a bug somewhere else in our layout.
For the present, just leave a TODO comment there.