persist [nfc]: Start type-checking purgeStoredState
In doing so, we move some types from index.js.flow to a new file,
types.js. We fix them to be real, non-unclear types. For the types
that aren't engaged in this file, we make them maximally permissive
to the code outside this library. When we get to other parts of this
library that use those types, we'll tighten up the types.