types [nfc]: Remove some redundant read-only-property annotations.
Each of these `+` marks means the property is read-only:
  https://flow.org/en/docs/types/interfaces/#toc-interface-property-variance-read-only-and-write-only

In each case that's the behavior we want.  But in each case we
already have a `$ReadOnly<…>` around the whole object type, which
applies a `+` to each property:
  https://flow.org/en/docs/types/utilities/#toc-readonly

So putting `+` on an individual property is redundant.  Take it out,
to avoid confusing people reading the codebase and trying to
understand what the various type annotations mean:
  https://github.com/zulip/zulip-mobile/pull/5039#discussion_r724636461

(In each of these cases, we added the `+` marks at a time when there
wasn't a `$ReadOnly` there, and then added the latter as part of a
broad automated refactoring.)