i18n [nfc]: Make the translations.js generation command locale-independent
This was producing different results for Chris from what it did for me;
it seems his shell expands the file glob in byte order, while mine uses
the locale's ordering (and we both have the locale `en_US.UTF-8`.)

It's useful for the order to be stable whoever runs the command.  Add a
sort, forcing bytewise ordering, to make it so.