sharing: Track a stream ID, as well as stream name
We don't yet actually do anything with this stream ID.  But with
this change we now have one available at the two points in this code
where we currently use the name to identify the stream: in sending
the message, and in navigating to the conversation.

In particular this takes care of one of the few remaining spots that
wasn't ready for narrows to start carrying stream IDs.