api: Adapt to API change in update-stream, conditionally
Until server feature level 64 earlier this year, the server API
expected these two string parameters to be JSON-encodings of the
desired strings, rather than simply the desired strings.  So the
`JSON.stringify` calls here were required for correct behavior
(even though the comments' reasoning, about "unsafe characters",
does not inspire confidence.)

But then the API was cleaned up to simply take the strings
directly, and broke compatibility as a result of accidentally
not correctly checking whether the mobile app was using this
feature.  This causes an edited stream name or description
to get the JSON encoding of the desired value (in particular,
to always get enclosed in double-quotes "…"), instead of the
actual desired value: #4747.

Fix that, conditioning on the server's feature level so that we
continue to behave correctly on older servers as well as new ones.

This includes the changes from #4748, which switched unconditionally
to the new API.

Includes-changes-by: Gautam Arora <gautamarora6248@gmail.com>
Fixes: #4747