ShareFileAndroid [nfc]: Fix a small import error
ShareFileAndroid.kt has nothing to do with
react-native-image-picker, so it's not obvious that it needs to
import from it. The import was likely the result of an auto-complete
gone slightly wrong.

Looking at the thing it's importing in the image-picker library,
it's actually not something that we'd ever have a good reason to
import anywhere. It's just a trivial subclass of
androidx.core.content.FileProvider.

So, just use androidx.core.content.FileProvider.