android: Switch over to "scoped storage"
As we understand it, the main work for this was done in #5060, when
we upgraded react-native-image-picker to a version that uses the
relevant APIs. We confirmed that image uploads broke if we tried to
do this before #5060, and that they don't break now, after #5060.

Here, we opt into scoped storage when we can, i.e., on Android 10
and above, by no longer requesting the "legacy" model with
`requestLegacyExternalStorage`.

We follow a guide doc [1] that says,

> If your app uses scoped storage, it should request storage-related
> permissions only for devices that run Android 9 (API level 28) or
> lower. You can apply this condition by adding the
> android:maxSdkVersion attribute to the permission declaration in
> your app's manifest file […]

That doc leaves out some helpful context that Greg explained in the
office today. While it's true that we'd need to continue requesting
READ_EXTERNAL_STORAGE on newer Android versions in order to query
for other apps' files *directly in our app's own code*, that doesn't
mean that our upload-files flows are broken without that permission.
Those flows delegate the user-chooses-file task to a separate
activity that's independent of Zulip, and trusted by Android to
clearly and safely perform the task. That activity has the
permissions it needs to get files from a common location that other
apps have saved files to.

[1] https://developer.android.com/training/data-storage/shared/media#scoped_storage_enabled

Fixes: #4283