flow [nfc]: Add an annotation to prevent an escaped generic
Flow v0.141.0 would give an escaped-generic error here, below.

Background story is in this blog post:
  https://medium.com/flow-type/flows-improved-handling-of-generic-types-b5909cc5e3c5

… but isn't really needed in order to fix the error: it comes with
a nice concrete recommendation to add a type annotation for this
function expression's return type.  The idea of this function is to
return an object of the same type as the one it got, so the intended
return type is `O`; say that.

(As the blog post says, Flow started giving these errors in v0.136,
before the version we're already on.  So I suppose this comes from
some bugfix that causes it to catch a few cases it didn't before.)

Error ------------------------------------ src/outbox/outboxReducer.js:35:54

Cannot spread object literal because type variable `O` [1] cannot escape
from the scope in which it was defined [2] (try adding a type annotation
to return [3]). [escaped-generic]

   src/outbox/outboxReducer.js:35:54
   35|         item.id !== action.local_message_id ? item : { ...(item: O), isSent: true },
                                                            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ [1]

References:
   src/outbox/outboxReducer.js:34:25
   34|       return state.map(<O: Outbox>(item: O) =>
                               ^ [2]
   src/outbox/outboxReducer.js:34:44
   34|       return state.map(<O: Outbox>(item: O) =>
                                                   [3]