android: Update target SDK version to 30, i.e. Android 11
This change will be required in order to upload new releases to the
Play Store, effective 2021-11-01, which is soon.

The change causes Android 11 and later to apply to our app a number
of behavior changes introduced in that version, documented here:
  https://developer.android.com/about/versions/11/behavior-changes-11

The most significant task we've had to do for this is to switch over
to "scoped storage", which is now enforced on Android 11 for apps
that target Android 11. We finished that in the previous commit.

See Greg's notes from reading that doc at
  https://github.com/zulip/zulip-mobile/issues/4284#issuecomment-947201084.

We've tested the following file-storage-related flows on an emulator
running Android 11:

- Uploading images from the camera or photo library, including a
  photo that was put in the library by a different app
- Uploading a PDF that was downloaded by a different app (Google
  Chrome)
- Image download from the lightbox
- Opening an uploaded PDF in a message

For "one-time permissions" [1], we should follow up by checking if
the effect of #3787, "First attempt to open camera/gallery fails",
is that the camera/gallery flows won't work at all. But this doesn't
seem to be related to the value at `targetSdkVersion`. One-time
permissions has been usable with apps running Android 11 regardless
of what apps choose for `targetSdkVersion`.

We do add some intent queries to our AndroidManifest.xml that
actually should have been there all along; see comments there. But
their absence (especially "android.intent.action.VIEW", which wasn't
even in the merged manifest) would have been much more noticeable
because of "package visibility filtering":
  https://developer.android.com/training/package-visibility

[1] https://developer.android.com/about/versions/11/privacy/permissions#one-time

Fixes: #4284