drawer: Add Drawer and drawer icon to top left
Added a DrawerNavigation.js file in src/main folder to hold the navigation of the Drawer.
Created a drawerContent.js file in src/drawerContent to hold the contents of the drawer.
Removed the icons in HomeScreen.js and sent them in the DrawerContent.js

Resolving conflicts on android/build.gradle

Removed the search icon from side drawer and sent it to the bottom navigation.
Made sure the back arrow on the searchMesages screen works appropriately.

Added the organization to the left of the Logo icon.
Labeled the star on the drawer to 'Starred messages' as required.

Added real_name to the redux store, so now we store the real_name.
Added real_name to the side drawer.
Also imported that state in the DrawerContent screen.