share-to: Stop mutating input value for mandatory-topics enforcement
In order to identify a topic as empty, we first have to take the raw
input and .trim() it, and then compare the result of that to the
empty string. That way, topics like " " and "   " will correctly be
treated as empty.

For the trimming part, we've just been calling .trim() every time
the input value changes and immediately setting the input value to
the output of that .trim().

This had the annoying side effect that it was really frustrating to
set a topic with multiple words in it, like "test topic". When you'd
typed as far as "test ", it would trim out the trailing space. So,
without a workaround, you'd end up with "testtopic". Then, if you
wanted, you could move the cursor back in between the words and
insert a space, to make "test topic". (That space wouldn't be
trimmed out because it's not at the beginning or end of the string.)

So, get rid of the side effect of changing the input value.