settings [nfc]: Rename an action as SET_GLOBAL_SETTINGS
This was called just SETTINGS_CHANGE.  We want to specifically
distinguish changing a global from a per-account setting, because
those will start being stored in different places.

The action already only touches global settings -- we don't have any
per-account settings that aren't server data, and therefore covered by
event actions instead -- so all it needs is a rename to make that
clear, and so that if we add a local per-account setting in the future
it'll be natural to make a new action type for that.