webview js: Drop remaining polyfills for old browsers
We've just bumped our minimum Android and therefore minimum Chrome
version, and it has both of these.  So does our minimum Safari version;
the comment on one says "at least Safari 13", but consulting MDN:
  https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/from#browser_compatibility

says it's been there since Safari on iOS 9.

And with that following fc0add55c, we have no polyfills left!
That might not always be true in the future, but it's nice to
have it true for now.