android notif [nfc]: Add TODO for apparent bug in clearing notifications
If the user reads the last messages in several conversations at once,
and those are their only active notifications -- for example, if they
have a couple of recent PMs in different threads, and visit the
"Private messages" view in the web app -- then it looks from the code
here like we'll delete those individual conversations' notifications,
but not the summary for the group, leaving behind a ghost header.

I noticed this just now from reading the code.  I haven't confirmed
empirically that this indeed happens.  If there's a subtle reason
it doesn't, though, we should resolve this TODO by making that clear.

[Added later: Indeed, confirmed this empirically; filed as #5119.]