android notif [nfc]: Fix a suppressed linter error
The linter is right about this -- if we're going to use the SDK's
NotificationChannel, we should be passing it a constant from the
SDK's NotificationManager, not from the AndroidX / Jetpack version
of these APIs.

When looking this bit of API up, I also discovered there's now an
AndroidX version of NotificationChannel.  Switching to it will
require upgrading dependencies, though; so leave that out of this
commit, and for the moment just leave a TODO comment.