android notif [nfc]: Inline several single-use variables
And bring others next to where they're used.

This makes it easier to read both the definition and the use-site
in the context of each other.

For `sender` and `selfUser`, having a name is helpful because it's
otherwise not immediately obvious what these values are meant to be
(e.g. as the third argument to `MessagingStyle#addMessage`.)  For the
locals passed to `setConversationTitle` and `setGroupConversation`,
though, those method names carry all the same information the
variable's name did.