android notif: Set FLAG_IMMUTABLE on our PendingIntent
This is what we want, because this isn't one of the unusual
cases where the PendingIntent is meant to have some of its
details overridden when it's turned into an actual Intent to
send.  See explanation in docs at FLAG_MUTABLE:
  https://developer.android.com/reference/kotlin/android/app/PendingIntent.html#flag_mutable

Android Studio had recently started giving me a rather inscrutable
warning about this.  Then I was reading the PendingIntent docs
in order to make other changes to this code, and ran across this
bit of background.

I tested opening a couple of notifications after making this change,
and everything worked as normal.