ComposeBox [nfc]: Stop using FloatingActionButton for submit button (12/n)
Stop redundantly setting width/height/overflow.

IconDone and IconSend are both square, so the width and height will
be computed to be the same as we've been setting explicitly here:

  Icon's margin   + Icon's width
  8 left, 8 right + 16           = 32

  (Similarly in the vertical direction.)

And with the explicit width and height gone, it doesn't make sense
to set `overflow` to anything.

Since this code came from FloatingActionButton, also go and put a
comment there about a potential bug if the caller passes an icon
that's wider than it is tall. None of its callers currently do that,
which is nice.