ComposeBox: Check mandatory-topic in getValidationErrors helper
The effect of this is that we now disable the submit button if a
mandatory topic is missing.

That means that the code in `handleSend` for checking that condition
isn't currently active, because `handleSend` can't be called when
the submit button is disabled.

Rather than deleting that code in `handleSend`, though, we plan to
make it so that `handleSend` still gets called on pressing the
submit button when the submit button is disabled. When we do that,
we'll make sure to handle all cases where
`this.getValidationErrors()` is non-empty, and give the user
actionable feedback on all the validation errors that apply. Like we
recently did in the share-to logic in 630a543de.