persist [nfc]: Cut always-null error property on rehydrate action
At the call site of this `rehydrateAction` action creator, the
`err` variable is always null anyway, so this wasn't carrying any
useful information.

(Locally it's clear that it's always falsy, because of the
`if (err)` conditional above.  From the implementation of
`getStoredState` I believe it's then always null... but even
if there is some odd path that would let it be `false` or 0
or whatever, that still wouldn't be very useful.)