persist [nfc]: Use one broader try/finally in persistStore
This simplifies things so that we only call `complete` in one place.

It also makes this code a lot more robust against failing to call
`complete`, and hence failing to resume the persistor.  There was a
stretch of code around `purgeKeys.forEach` that wasn't covered by
the try/finally; if that code threw an exception for some reason,
the persistor would remain paused and never be resumed.

An exception in that bit of code would be a bug, and on inspection I
*think* we didn't have any such bug, which means this change is NFC.
But given that gap in the try/finally coverage, it would have been
easy for some future change to introduce something there that did
sometimes throw an exception, and then we'd have had a live bug.