deps: Add expo-sqlite
This isn't quite the latest version; for now we'll use v9.x in order
to stick with "unimodules".  Issue #5133 is open for the migration
that'll let us advance to v10.x.

Include a libdef, generated with flowgen with small manual fixups
as described in comments.  The libdef is actually from a newer
version because that's what I tried first; I tried rerunning flowgen
on the older version, and the differences in the new one are pure
compatible improvements (cleaning up some type definitions; adding
comments; replacing `any[]` as the type for `args` on `executeSql`.)