android notif [nfc]: Make intent URL clearly different from web-auth URLs
The URLs we get in the web-auth flow (parsed in src/start/webAuth.js)
look like `zulip://login?…` -- they have not only a scheme of "zulip"
but an authority of "login".  ("Authority" being the generic URL term
for what in schemes like "https" is the host, port, and so on.)

We'd had these URLs look like `zulip:…`, filled in with the message
key.  These couldn't actually collide, because the message key never
starts with `//`… but it'd be nicer to have them cleanly namespaced.
Do that by giving them their own, distinct, authority.