keyboard-avoiding ios: Begin forking RN's KeyboardAvoidingView
From
  node_modules/react-native/Libraries/Components/Keyboard/KeyboardAvoidingView.js
with v0.64.2 of `react-native`, adjusting only the license pointer
in the copyright header.

We'll make the fork usable soon, not in this pure copy-paste commit.

We don't like this implementation (see, e.g., 70eca0716, which took
a lot of painstaking investigation), but we want to fix #5162 as
soon as we can.

For a not-quite-perfect attempt at reimplementing it from scratch in
our style (and with #5162 fixed), see my branch
`reimplement-keyboard-avoiding-view` on GitHub.

We can't easily make a jank-free solution in a normal React Native
way because we can't have perfect information about the layout on
every frame. React Native exposes it to us by consuming event
listeners and providing asynchronous query functions, and we end up
having to learn about different aspects of the layout at different
times.

The best hope for a jank-free solution is to use native iOS APIs. If
we're feeling adventurous and we find the time, we should really try
hard to make React Native play along with iOS's
`keyboardLayoutGuide` API (iOS 15+).

The first four minutes of this video should be really useful
background on the `keyboardLayoutGuide` feature:
  https://developer.apple.com/videos/play/wwdc2021/10259/

It works within iOS's "Auto Layout" system:
  https://developer.apple.com/library/archive/documentation/UserExperience/Conceptual/AutolayoutPG/index.html

But then we'd have to go and

(a) Make a native component:
      https://reactnative.dev/docs/native-components-ios

(b) Figure out how to make React Native's propagate-from-JavaScript
    layout system (Yoga) not fight with the native iOS layout
    system, including "Auto Layout" and `keyboardLayoutGuide`.
    Possibly we can pick up some clues from
    `react-native-safe-area-context` for this?

From another angle, since the jank is regularly seen on screen
orientation changes between portrait and landscape, we could
consider just unsupporting the landscape orientation. On very common
device types, like phones, it's just not as easy to use the app in,
especially for composing messages.

(cherry picked from commit d7d91dcfb63eae31cc39526337649743699da6a7)