i18n [nfc]: More concisely create styled, translated text
Rendering a translated string in the UI with styling applied is the
job of `Label`. Let `Label` do its job in these two places.

`Label` (because of `RawLabel`) would normally apply some default
styles, in the `color` and `fontSize` attributes. Both callers
happen to override both attributes, so this is NFC.