wip sqlite add tests -- one failing, on transaction
We could work around this for the case when the await is actually of a
query result -- rework everything into callback style, so the next bit
of the transaction runs immediately in that result callback, before
the transaction engine checks its queue again.  Ugly but it'd work.

But then we also have cases where something else is async, like
compression.  I think the underlying expo-sqlite (and really the
websql implementation underneath it) may not make it possible to
handle that.

Then on top of that, and rather unforgivably: the failure is
completely silent!  We go and call `tx.executeSql` on a transaction
that's already decided it's complete... and there's no error.  It just
pushes the new task on the queue, which it never again intends to
check.  This would be so easy to fix, too: just a check for
`self._complete` at the top of `executeSql` (in
`lib/websql/WebSQLTransaction.js` in `@expo/websql`), throwing an
error when that's true.