expo-sqlite tests: Basic transactions tests
These just exercise simple happy cases of using transactions.

If you do something more complicated -- like the transaction's code
hits an unexpected error midway through, or you need to do some
asynchronous work with the result of one query before making the
next -- then expo-sqlite turns out to not behave so well.

Those tests get more complicated to write (partly because of the
bad behavior itself, and the need to protect the Jest environment
from it), so we'll add them separately.  We'll also work around
them in the promisified wrapper we'll be adding over expo-sqlite.