migrations: Fix buggy old migration 22, on PM narrow keys
Fixing this will have very little practical effect; the original,
buggy, migration was released nearly a year ago, so almost all
users that were going to run it will have already run it by now.

But now that we've written tests for these migrations and we know
it's buggy, it's pretty easy to fix.  And doing so lets us clean up
the test of the broken behavior, and also record our assessment of
the consequences of the bug.  (Which aren't especially bad,
fortunately.)