msglist tests: Have update-content tests simulate DOM manipulation
We use JSDOM, a headless browser implementation. It ships with Jest,
and Jest recommends it for testing DOM manipulation:
  https://jestjs.io/docs/26.x/tutorial-jquery

This confirms nicely (as of course we've seen all along in
production) that the HTML strings we generate are indeed parseable
into DOM nodes. If they weren't, we would see more than these
fine/boring whitespace changes in the snapshots. So this is a good
thing to be doing in any case.

But also, we'd like to reuse most of the code in this file to test
the new edit-sequence inbound event type (for #3851). The DOM
manipulation code for those events isn't as easy to verify as the
corresponding code for these "content" events, which is basically a
one-liner:

  msglistElementsDiv.innerHTML = uevent.content;

So, simulating DOM manipulation will be necessary for our peace of
mind when we get there. Doing it now helpfully normalizes away some
whitespace quirks in the snapshots. We'll be reusing the snapshots
entirely, to ease the transition.